+++ lib/axios.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+const instance = axios.create({
+  baseURL: 'https://bootcamp-api.codeit.kr/api/'
+})
+
+export default instance;
