+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^1.6.5",
     "mongoose": "^8.0.4",
     "next": "13.5.6",
     "react": "^18",
