+++ pages/signin/index.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import JoinSns from '@/src/components/joinSns/JoinSns';
 import JoinLink from '@/src/components/joinLink/JoinLink';
 import styled from 'styled-components';
-import SigninForm from '@/src/components/userInput/SigninForm';
+import Signinform from '@/src/components/userInput/Signinform';
 
 const Container = styled.div`
   display: flex;
@@ -25,7 +25,7 @@ export default function SigninPage() {
       <Container>
         <Wrap>
           <JoinLink />
-          <SigninForm />
+          <Signinform />
           <JoinSns />
         </Wrap>
       </Container>
