+++ src/components/userInput/SignupForm.tsx
@@ -1,0 +1,168 @@
+import { useEffect, useState } from 'react';
+import styled from 'styled-components';
+import { BASE_URL } from '@/constant';
+import { useForm } from 'react-hook-form';
+import UserButton from '@/src/components/userButton/UserButton';
+import { InputValue, Error } from '@/src/components/userInput/SignFormProvider';
+
+async function onSubmit(USER_INFO: InputValue, setError: Error) {
+  try {
+    const response = await fetch(`${BASE_URL}/sign-up`, {
+      method: 'POST',
+      body: JSON.stringify(USER_INFO),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+
+    const { data } = await response.json();
+    localStorage.setItem('accessToken', data.accessToken);
+    if (localStorage.getItem('accessToken')) window.location.href = '/folder';
+
+    if (!response.ok) throw new Error('이미 가입된 회원입니다.');
+  } catch {
+    return;
+  }
+}
+
+async function checkEmail(USER_INFO: InputValue, setError: Error) {
+  try {
+    const response = await fetch(`${BASE_URL}/sign-up`, {
+      method: 'POST',
+      body: JSON.stringify(USER_INFO.email),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+
+    const { data } = await response.json();
+    localStorage.setItem('accessToken', data.accessToken);
+    if (localStorage.getItem('accessToken')) window.location.href = '/folder';
+
+    if (!response.ok) throw new Error('중복된 이메일입니다.');
+  } catch {
+    setError('email', { message: '이미 사용중인 이메일입니다' });
+  }
+}
+
+export default function SignupForm() {
+  const {
+    register,
+    handleSubmit,
+    setError,
+    trigger,
+    formState: { errors },
+    getValues,
+  } = useForm<InputValue>({ mode: 'onBlur' });
+
+  const [togglePassword, setTogglePassword] = useState<boolean>(false);
+  const [togglePasswordCheck, setTogglePasswordCheck] =
+    useState<boolean>(false);
+
+  const handleClickPasswordCheck = () =>
+    setTogglePasswordCheck(!togglePasswordCheck);
+
+  const USER_INFO = {
+    email: getValues('email'),
+    password: getValues('password'),
+  };
+
+  useEffect(() => {
+    setTogglePassword(true);
+  }, []);
+
+  return (
+    <form onSubmit={handleSubmit(() => onSubmit(USER_INFO, setError))}>
+      <InputContainer>
+        <InputBox>
+          <Label htmlFor="signup-check-password">비밀번호 확인</Label>
+          <PassWord>
+            <Input
+              type={togglePassword ? 'text' : 'password'}
+              id="signup-check-password"
+              className={errors.passwordCheck ? 'active' : ''}
+              placeholder="비밀번호와 일치하는 값을 입력해 주세요"
+              {...register('passwordCheck', {
+                validate: {
+                  check: (val) => {
+                    if (getValues('password') !== val) {
+                      return '비밀번호가 일치하지 않습니다.';
+                    }
+                  },
+                },
+              })}
+              name="passwordCheck"
+            />
+            <EyeImg
+              src={
+                togglePasswordCheck
+                  ? '/image/ico-eye-on.svg'
+                  : '/image/ico-eye-off.svg'
+              }
+              alt={togglePasswordCheck ? '비밀번호 표시' : '비밀번호 숨기기'}
+              onClick={handleClickPasswordCheck}
+            />
+          </PassWord>
+          {errors.passwordCheck && (
+            <Messages>{errors.passwordCheck?.message}</Messages>
+          )}
+        </InputBox>
+      </InputContainer>
+      <UserButton />
+    </form>
+  );
+}
+
+const InputContainer = styled.div`
+  margin: 3rem 0 0;
+`;
+
+const InputBox = styled.div`
+  margin: 2.4rem 0 0;
+`;
+
+const Label = styled.label`
+  display: block;
+  margin: 0 0 1.2rem 0;
+  font-size: 1.4rem;
+  line-height: 1.6rem;
+`;
+
+const Input = styled.input`
+  outline: none;
+  border: none;
+  width: 100%;
+  padding: 1.8rem 1.5rem;
+  font-size: 1.6rem;
+  color: var(--gray100);
+  border-radius: 0.8rem;
+  border: 1px solid var(--gray20);
+  background-color: var(--white);
+
+  &:focus {
+    border: 1px solid var(--primary);
+  }
+
+  &.active {
+    border: 1px solid var(--red);
+  }
+`;
+
+const Messages = styled.p`
+  margin: 0.6rem 0 0;
+  color: var(--red);
+  font-size: 1.4rem;
+  line-height: 1.6rem;
+`;
+
+const PassWord = styled.div`
+  position: relative;
+`;
+
+const EyeImg = styled.img`
+  position: absolute;
+  top: 50%;
+  right: 1.5rem;
+  transform: translateY(-50%);
+  cursor: pointer;
+`;
